move repository to common module
