+++ shared_database/services/customer_service/src/main/java/nawaphon/microservices/customer_service/services/MainService.java
@@ -2,7 +2,7 @@ package nawaphon.microservices.customer_service.services;
 
 import nawaphon.microservice.main.common.pojo.Customer;
 import nawaphon.microservice.main.common.pojo.ResponseMessage;
-import nawaphon.microservices.customer_service.repositories.CustomerRepository;
+import nawaphon.microservice.shared_database.common.repositories.CustomerRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Example;
