+++ shared_database/services/order_service/src/main/java/nawaphon/microservices/order_service/services/NonRequiredTransactionalService.java
@@ -3,7 +3,7 @@ package nawaphon.microservices.order_service.services;
 import nawaphon.microservice.main.common.pojo.Customer;
 import nawaphon.microservice.main.common.pojo.Order;
 import nawaphon.microservice.main.common.pojo.ResponseMessage;
-import nawaphon.microservices.order_service.repositories.OrderRepository;
+import nawaphon.microservice.shared_database.common.repositories.OrderRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Example;
